App is working after refactor. Prepare for code cleanini
