+++ app/src/main/AndroidManifest.xml
@@ -15,21 +15,15 @@
         android:supportsRtl="true"
         android:theme="@style/Base.Theme.PlaylistMaker"
         tools:targetApi="31">
-        <activity
-            android:name=".medialibrary.ui.activity.MediaLibraryActivity"
-            android:exported="false" />
+
         <activity
             android:name=".player.ui.PlayerActivity"
             android:exported="false" />
         <activity
-            android:name=".search.ui.SearchActivity"
-            android:exported="false" />
-        <activity
-            android:name=".settings.ui.SettingsActivity"
-            android:exported="false" />
-        <activity
-            android:name=".main.ui.MainActivity"
-            android:exported="true">
+            android:name=".main.ui.RootActivity"
+            android:exported="true"
+            android:windowSoftInputMode="adjustNothing">
+
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
