+++ app/src/main/java/com/practicum/playlistmaker/main/ui/RootActivity.kt
@@ -1,0 +1,56 @@
+package com.practicum.playlistmaker.main.ui
+
+import android.os.Build
+import android.os.Bundle
+import android.view.LayoutInflater
+import androidx.activity.enableEdgeToEdge
+import androidx.appcompat.app.AppCompatActivity
+import androidx.core.view.ViewCompat
+import androidx.core.view.WindowInsetsCompat
+import androidx.core.view.isVisible
+import androidx.navigation.fragment.NavHostFragment
+import androidx.navigation.fragment.findNavController
+import androidx.navigation.ui.setupWithNavController
+import com.google.android.material.bottomnavigation.BottomNavigationView
+import com.practicum.playlistmaker.R
+import com.practicum.playlistmaker.databinding.RootActivityBinding
+
+class RootActivity : AppCompatActivity() {
+
+    private lateinit var binding: RootActivityBinding
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        enableEdgeToEdge()
+        binding = RootActivityBinding.inflate(layoutInflater)
+        setContentView(binding.root)
+
+        ViewCompat.setOnApplyWindowInsetsListener(binding.root) { v, insets ->
+            val systemBars = insets.getInsets(WindowInsetsCompat.Type.systemBars())
+            v.setPadding(systemBars.left, systemBars.top, systemBars.right, systemBars.bottom)
+            insets
+        }
+
+        val navHostFragment =
+            supportFragmentManager.findFragmentById(R.id.rootContainerView) as NavHostFragment
+        val navHostController = navHostFragment.findNavController()
+
+        val bottomNavigationView = binding.bottomNavigationView
+        bottomNavigationView.setupWithNavController(navHostController)
+
+        navHostController.addOnDestinationChangedListener { _, destination, _ ->
+            when (destination.id) {
+                R.id.searchFragment -> {
+                    bottomNavigationView.isVisible = false
+                    binding.shadowView.isVisible = false
+
+                }
+                else -> bottomNavigationView.isVisible = true
+            }
+
+        }
+    }
+}
+
+
+
