+++ app/src/main/java/com/practicum/playlistmaker/medialibrary/ui/activity/MediaLibraryFragment.kt
@@ -27,11 +27,6 @@ class MediaLibraryFragment : Fragment() {
         savedInstanceState: Bundle?
     ): View? {
         _binding = MedialibraryFragmentBinding.inflate(inflater, container, false)
-        ViewCompat.setOnApplyWindowInsetsListener(binding.root) { v, insets ->
-            val systemBars = insets.getInsets(WindowInsetsCompat.Type.systemBars())
-            v.setPadding(systemBars.left, systemBars.top, systemBars.right, systemBars.bottom)
-            insets
-        }
         return binding.root
     }
