+++ app/src/main/res/color/nav_item_color_night.xml
@@ -1,0 +1,5 @@
+<?xml version="1.0" encoding="utf-8"?>
+<selector xmlns:android="http://schemas.android.com/apk/res/android">
+    <item android:color="@color/yp_blue" android:state_checked="true" />
+    <item android:color="@color/white" android:state_checked="false" />
+</selector>
