+++ app/src/main/res/layout/root_activity.xml
@@ -5,10 +5,11 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".main.ui.MainActivity">
+    style="@style/MainActivityLayoutStyle"
+    tools:context=".main.ui.RootActivity">
 
 <androidx.fragment.app.FragmentContainerView
-    android:id="@+id/fragmentContainerView"
+    android:id="@+id/rootContainerView"
     android:layout_width="0dp"
     android:layout_height="0dp"
     android:name="androidx.navigation.fragment.NavHostFragment"
@@ -20,16 +21,23 @@
     app:layout_constraintStart_toStartOf="parent"
     app:layout_constraintTop_toTopOf="parent"/>
 
+    <View
+        android:id="@+id/shadowView"
+        android:layout_width="match_parent"
+        android:layout_height="1dp"
+        android:background="@color/grey_pale"
+        app:layout_constraintBottom_toTopOf="@id/bottomNavigationView"/>
+
     <com.google.android.material.bottomnavigation.BottomNavigationView
         android:id="@+id/bottomNavigationView"
+        style="@style/BottomNavigationStyle"
         android:layout_width="0dp"
         android:layout_height="wrap_content"
 
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.0"
-
         app:layout_constraintStart_toStartOf="parent"
+
         app:menu="@menu/bottom_navigation_menu" />
