+++ app/src/main/res/layout/settings_fragment.xml
@@ -5,24 +5,32 @@
     style="@style/InnerActivityLayoutStyle"
     tools:context=".settings.ui.SettingsFragment">
 
-    <FrameLayout
+    <Toolbar
+        style="@style/ToolBarStyle"
+        android:id="@+id/toolBar"
         android:layout_width="match_parent"
-        android:layout_height="56dp">
+        android:layout_height="wrap_content"
+        android:title="@string/settings"
+        android:titleTextAppearance="@style/ToolbarHeader"/>
 
-        <FrameLayout style="@style/InnerActivityHeaderFrameLayout" />
+<!--    <FrameLayout-->
+<!--        android:layout_width="match_parent"-->
+<!--        android:layout_height="56dp">-->
 
-        <TextView style="@style/TitleTextStyle"
-            android:layout_marginTop="16dp"
-            android:layout_marginLeft="52dp"
-            android:layout_marginRight="100dp"
-            android:text="@string/settings" />
+<!--        <FrameLayout style="@style/InnerActivityHeaderFrameLayout" />-->
 
-        <Button style="@style/SettingsActivityButtonBack"
-            android:id="@+id/buttonSettingsBack"
-            android:layout_gravity="start"
-            app:icon="@drawable/vector_arrow_back"
-            app:iconGravity="textStart"/>
-    </FrameLayout>
+<!--        <TextView style="@style/TitleTextStyle"-->
+<!--            android:layout_marginTop="16dp"-->
+<!--            android:layout_marginLeft="52dp"-->
+<!--            android:layout_marginRight="100dp"-->
+<!--            android:text="@string/settings" />-->
+
+<!--        <Button style="@style/SettingsActivityButtonBack"-->
+<!--            android:id="@+id/buttonSettingsBack"-->
+<!--            android:layout_gravity="start"-->
+<!--            app:icon="@drawable/vector_arrow_back"-->
+<!--            app:iconGravity="textStart"/>-->
+<!--    </FrameLayout>-->
 
     <LinearLayout
         android:layout_width="match_parent"
