+++ app/src/main/res/menu/bottom_navigation_menu.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <menu xmlns:android="http://schemas.android.com/apk/res/android">
     <item
-        android:id="@+id/SearchFragment"
+        android:id="@+id/searchFragment"
         android:icon="@drawable/vector_search"
         android:title="@string/search"/>
