+++ app/src/main/res/values-night/styles.xml
@@ -1,11 +1,15 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
 
+<!--    <style name="MainActivityLayoutStyle">-->
+<!--        <item name="android:layout_width">match_parent</item>-->
+<!--        <item name="android:layout_height">match_parent</item>-->
+<!--        <item name="android:background">@color/yp_black</item>-->
+<!--        <item name="android:orientation">vertical</item>-->
+<!--    </style>-->
+
     <style name="MainActivityLayoutStyle">
-        <item name="android:layout_width">match_parent</item>
-        <item name="android:layout_height">match_parent</item>
         <item name="android:background">@color/yp_black</item>
-        <item name="android:orientation">vertical</item>
     </style>
 
     <style name="TitleTextStyle">
@@ -198,9 +202,23 @@
     </style>
 
     <style name="ToolBarStyle">
-        <item name="android:navigationIcon">@drawable/vector_player_button_back_black_mode</item>
         <item name="android:titleTextColor">@color/white</item>
         <item name="backgroundColor">@color/yp_black</item>
+    </style>
+
+
+    <style name="BottomNavigationStyle">
+        <item name="itemTextColor">@color/nav_item_color_night</item>
+        <item name="itemIconTint">@color/nav_item_color_night</item>
+
+        <item name="itemTextAppearanceActive">@style/BottomNavTextAppearance</item>
+        <item name="itemTextAppearanceInactive">@style/BottomNavTextAppearance</item>
+    </style>
+
+
+    <style name="BottomNavTextAppearance">
+        <item name="android:fontFamily">@font/ys_display_regular</item>
+        <item name="android:textSize">12sp</item>
     </style>
 
 </resources>
