+++ app/src/main/res/values/colors.xml
@@ -12,4 +12,5 @@
     <color name="grey">#AEAFB4</color>
     <color name="grey_pale">#E6E8EB</color>
     <color name="yp_black">#1A1B22</color>
+
 </resources>
