+++ app/src/main/res/values/styles.xml
@@ -11,12 +11,12 @@
         <item name="android:textFontWeight">500</item>
     </style>
 
-    <style name="MainActivityLayoutStyle">
-        <item name="android:layout_width">match_parent</item>
-        <item name="android:layout_height">match_parent</item>
-        <item name="android:background">@color/yp_blue</item>
-        <item name="android:orientation">vertical</item>
-    </style>
+<!--<style name="MainActivityLayoutStyle">&ndash;&gt;-->
+<!--       <item name="android:layout_width">match_parent</item>-->
+<!--      <item name="android:layout_height">match_parent</item>-->
+<!--      <item name="android:background">@color/white</item>-->
+<!--       <item name="android:orientation">vertical</item>-->
+<!--</style>-->
 
     <style name="MainActivityButtonStyle" parent="Widget.MaterialComponents.Button.UnelevatedButton">
         <item name="android:layout_width">match_parent</item>
@@ -212,9 +212,20 @@
 
 
     <style name="ToolBarStyle">
-        <item name="android:navigationIcon">@drawable/vector_arrow_back</item>
         <item name="android:titleTextColor">@color/yp_black</item>
     </style>
 
+    <style name="BottomNavigationStyle">
+        <item name="itemTextColor">@color/nav_item_color</item>
+        <item name="itemIconTint">@color/nav_item_color</item>
 
+        <item name="itemTextAppearanceActive">@style/BottomNavTextAppearance</item>
+        <item name="itemTextAppearanceInactive">@style/BottomNavTextAppearance</item>
+
+    </style>
+
+    <style name="BottomNavTextAppearance" >
+        <item name="android:fontFamily">@font/ys_display_regular</item>
+        <item name="android:textSize">12sp</item>
+    </style>
     </resources>
